+++ src/rocktick/resources/jobs.py
@@ -20,8 +20,8 @@ from .._response import (
 from ..pagination import SyncCursorPage, AsyncCursorPage
 from .._base_client import AsyncPaginator, make_request_options
 from ..types.one_off_job import OneOffJob
-from ..types.request_param import RequestParam
 from ..types.job_list_response import JobListResponse
+from ..types.http_request_param import HTTPRequestParam
 
 __all__ = ["JobsResource", "AsyncJobsResource"]
 
@@ -50,7 +50,7 @@ class JobsResource(SyncAPIResource):
         self,
         *,
         execute_at: int,
-        request: RequestParam,
+        request: HTTPRequestParam,
         max_response_bytes: Optional[int] | Omit = omit,
         max_retries: Optional[int] | Omit = omit,
         region: Optional[str] | Omit = omit,
@@ -130,7 +130,7 @@ class JobsResource(SyncAPIResource):
         max_response_bytes: Optional[int] | Omit = omit,
         max_retries: Optional[int] | Omit = omit,
         region: Optional[str] | Omit = omit,
-        request: Optional[RequestParam] | Omit = omit,
+        request: Optional[HTTPRequestParam] | Omit = omit,
         timeout_ms: Optional[int] | Omit = omit,
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.
@@ -236,7 +236,7 @@ class AsyncJobsResource(AsyncAPIResource):
         self,
         *,
         execute_at: int,
-        request: RequestParam,
+        request: HTTPRequestParam,
         max_response_bytes: Optional[int] | Omit = omit,
         max_retries: Optional[int] | Omit = omit,
         region: Optional[str] | Omit = omit,
@@ -316,7 +316,7 @@ class AsyncJobsResource(AsyncAPIResource):
         max_response_bytes: Optional[int] | Omit = omit,
         max_retries: Optional[int] | Omit = omit,
         region: Optional[str] | Omit = omit,
-        request: Optional[RequestParam] | Omit = omit,
+        request: Optional[HTTPRequestParam] | Omit = omit,
         timeout_ms: Optional[int] | Omit = omit,
         # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
         # The extra values given here take precedence over values defined on the client or passed to this method.