+++ src/rocktick/types/one_off_job.py
@@ -2,9 +2,9 @@
 
 from typing import List, Optional
 
-from .request import Request
 from .._models import BaseModel
 from .execution import Execution
+from .http_request import HTTPRequest
 
 __all__ = ["OneOffJob"]
 
@@ -20,7 +20,7 @@ class OneOffJob(BaseModel):
 
     region: str
 
-    request: Request
+    request: HTTPRequest
 
     max_response_bytes: Optional[int] = None